+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 调整移动端侧边面板树缩进 | Adjust the indentation of the tree in the sidebar panel on the mobile client.
 - 调整已完成任务列表项样式 | Adjust the style of the completed task list item.
 - 使用 `Less` 重构 `theme.css` | Refactored `theme.css` with `Less`.
+- 新增 `分享` 块菜单项 | Add block menu item `Share`.
 
 ## v1.7.6/2023-04-05